requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
dj-database-url==1.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dlib==19.24.8
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
gunicorn==23.0.0
idna==3.10
imageio==2.37.0
imutil==0.3.4
imutils==0.5.4
joblib==1.5.1
kombu==5.5.4
lazy_loader==0.4
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tifffile==2025.6.11
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
